Create line limits using a linear model trained on the UK grid

These functions set the line limits of network using a model trained on the UK power grid

Line_Limit_Volt_PF(g, PowerFlow, Voltage, Line_Limit = "Link.Limit")

Line_Limit_PF(g, PowerFlow, Line_Limit = "Link.Limit")

Arguments

g
An igraph object representing a power grid. must have a power flow attribute.

PowerFlow
The name of the graph attribute which contains the power flow over each line.

Voltage
The name of the graph attribute which contains the Voltage. The attribute should be numeric.

Line_Limit
The attribute name that will be used when setting the line limits.

Value

The original network with an additional edge attribute that is a numeric line limit created by; Line Limit = \(a*PowerFlow + b*Voltage + c\)

Details

The models used here were trained on a network representing a simplified version UK high voltage network similar to that used in the Electricity Ten Year Statement (ETYS) dataset. The model was validated using ten fold cross validation. In testing the Volt PF was a top performer or the best performing model in all tests performed in the paper Bourne et al. 2019. The PF model generally performed well but not at the level of the Volt_PF model. The Volt_PF model was trained using three difference voltage levels 132, 275 and 400Kv. The relationship between Voltage and log10 line limits appeared linear when correcting for power flow. However, we cannot say what will happen to the relationship at very high or very low voltage levels. The reccomended order in which artificial line limits should be used given available data is as follows.

Line_Limit_Volt_PF

Line_Limit_PF

Proportional_Load

See also

Examples

#Creat line limits using both methods
test_g_Volt <- Line_Limit_Volt_PF(g, "PowerFlow", "Voltage", Line_Limit = "Line_Limit_VoltPF")

#> Error in "igraph" %in% class(graph): object 'g' not found

test_g_PF <- Line_Limit_PF(g, "PowerFlow", Line_Limit = "Line_Limit_PF")

#> Error in "igraph" %in% class(graph): object 'g' not found

#Compare the resulting line limits.
tibble(VoltPF = get.edge.attribute(test_g_Volt, "Line_Limit"),
    PF =  get.edge.attribute(test_g_PF, "Line_Limit"))

#> Error in tibble(VoltPF = get.edge.attribute(test_g_Volt, "Line_Limit"),     PF = get.edge.attribute(test_g_PF, "Line_Limit")): could not find function "tibble"

#There can be substantial difference between the two due to the additional information
#provided by the voltage