Create proportional loading line limits

This function sets the line limits to the tolerance value \(\alpha\) chosen by the user.

Proportional_Load(g, alpha, PowerFlow = "PowerFlow", Link.Limit = "Link.Limit")

Arguments

g
An igraph object representing a power grid. must have a power flow attribute.

alpha
The system tolerance. A number larger than 1.

PowerFlow
The name of the graph attribute which contains the power flow over each line.

Link.Limit
The attribute name that will be used when setting the line limits.

Value

The original network with an additional edge attribute that is a numeric line limit created by; Line Limit = \(\alpha*PowerFlow\)

Details

Proportional loading should only be used when there are no other options available as it can be challenging to set the value of \(\alpha\) close to the system mean. Bourne et al. 2019 find that values of \(\alpha\) that deviate substantially from the actual value of \(\alpha\) given that load profile will not represent the behaviour of the power grid under attack or collapse.

Examples

Proportional_Load(g, alpha = 1.05)

#> Error in "igraph" %in% class(graph): object 'g' not found